requirements_gui.txt
fastapi
uvicorn
python-multipart
pyjwt
requests
